datetime api mods merged from 2.5 branch
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@795 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c/nitf/include/nitf/DateTime.h
@@ -118,7 +118,7 @@ NITFAPI(NITF_BOOL) nitf_DateTime_setMinute(nitf_DateTime *dateTime,
  * Sets the second of a DateTime object and updates the timeInMillis.
  */
 NITFAPI(NITF_BOOL) nitf_DateTime_setSecond(nitf_DateTime *dateTime,
-                                           int second,
+                                           double second,
                                            nitf_Error *error);
 
 /*!
